requirements.txt
fastapi==0.111.1
pydantic==2.9.2
uvicorn==0.23.2
youtube-transcript-api==0.6.0
langchain==0.1.0  # Latest stable version
langchain-community==0.0.142
langchain-google-genai==0.0.24
faiss-cpu==1.7.4
tiktoken==0.4.0
python-dotenv==1.0.0
